chore: split release-please action
Run the two release-please commands, release-pr and github-release, in
two different jobs with different tokens. The job for release-pr uses
teh default token, so that the PR is created with the github-actions
user. The github-release job is run with a personal access token to
ensure that further workflows can be automatically executed after an
release was created.